Edited the following parts:
Removed the Pin class to deal with the passwords as string.
Replaced the comperator class with a new one.
The comperator checks for length, if the lenght is the same, it checks
for the number size. 
The class has a exception for letter input.

removed unused code.
removed global variables, which can be local.
cleaned up the code.